backend/requirements.txt
google-adk>=0.3.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
pydantic>=2.0.0
google-generativeai>=0.8.0
google-cloud-aiplatform>=1.76.0
httpx>=0.27.0
cors>=1.0.1
